add plotly functionality
plots are now interactive. This is causing a problem where plotly refuses to acknowledge labels I had manually specified in my ggplot calls. Docs suggest it only cares about data and aesthetic mappings, so any re-naming should happen in processing. I am going to take the L on this one and submit with bad labels. Still, a task for the future!